Reorganise
Create freeFunction folders in the src and tests directories to better
organise the code.
