+++ src/freeFunctions/drawObjects.cpp
@@ -1,8 +1,8 @@
-#include "../include/wrappers/iRenderWindow.hpp"
-#include "../include/models/iLaserCannon.hpp"
-#include "../include/models/iLaser.hpp"
-#include "../include/models/iInvader.hpp"
 #include <vector>
+#include "../../include/wrappers/iRenderWindow.hpp"
+#include "../../include/models/iLaserCannon.hpp"
+#include "../../include/models/iLaser.hpp"
+#include "../../include/models/iInvader.hpp"
 
 void drawObjects(IRenderWindow &window, const ILaserCannon &cannon, const ILaser &laser, const std::vector<std::vector<IInvader *>> &invaders)
 {
