+++ src/freeFunctions/moveLaserCannon.cpp
@@ -1,4 +1,4 @@
-#include "../include/models/iLaserCannon.hpp"
+#include "../../include/models/iLaserCannon.hpp"
 
 void moveLaserCannon(ILaserCannon &cannon, float x)
 {
